moved tab manager to ui/tab_manager.hpp
